Update cluster type to only contain node pools count
Currently the cluster type contains a list of node pools with a full
schema making the output very verbose and inconsisent with other
operations.

This removes the node pools from the cluster types and replaces it with
a simple node pools count integer. To list details of node pools in a
cluster the list cluster node pools operation should be used.