Fix inconsistent template name in cluster options type
Currently there is a property named `cluster_template` in the cluster
options type. This property is used to optionally specify the name of
which cluster template to use when creating the cluster. Several other
types use the pattern of the property containing the name suffix to
indicated that it refers to the name of another resource.

This updates the cluster template name property to use the explicit name
suffix for consistency.